+++ lib/src/data/data_source/local/hive_entity/channel_hive_entity_20.dart
@@ -10,6 +10,10 @@ class ChannelHiveEntity extends HiveObject{
   @HiveField(0)
   String? channelId;
 
+  /// Default SubChannel Id
+  @HiveField(0)
+  String? defaultSubChannelId;
+
   /// Is Distinct
   @HiveField(1)
   bool? isDistinct;