+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.53.0
+* Feat: Expose default subchannelId for channel object
+* Deprecate: Channel creation with custom channelId
+
 ## 0.52.1
 * Fix: Exclude deleted posts from global feed and custom post ranking Live Collection